+++ enterprise/cmd/executor/internal/worker/worker.go
@@ -60,6 +60,9 @@ type Options struct {
 	// FilesOptions configures the client that interacts with the files API.
 	FilesOptions apiclient.BaseClientOptions
 
+	// DockerOptions configures the behavior of docker container creation.
+	DockerOptions command.DockerOptions
+
 	// FirecrackerOptions configures the behavior of Firecracker virtual machine creation.
 	FirecrackerOptions command.FirecrackerOptions